refactor: move tooltip
