+++ packages/components/tooltip/src/tooltip.js
@@ -6,8 +6,8 @@ import styled from '@emotion/styled';
 import isNil from 'lodash/isNil';
 import usePopper from 'use-popper';
 import { useFieldId, useToggleState } from '@commercetools-uikit/hooks';
-import createSequentialId from '../../utils/create-sequential-id';
-import SafeHTMLElement from '../../utils/helpers/safeHTMLElement';
+import createSequentialId from '../../../../src/utils/create-sequential-id';
+import SafeHTMLElement from '../../../../src/utils/helpers/safeHTMLElement';
 import { Body, getBodyStyles } from './tooltip.styles';
 
 const sequentialId = createSequentialId('tooltip-');
